+++ NEWS.md
@@ -34,6 +34,11 @@
 
 ## Features and bugfixes
 
+* External backtraces in error chains are now separately displayed (#1098).
+
+* Trace capture now better handles wrappers of calling handler in case
+  of rethrown chained errors.
+
 * The `print()` method of rlang errors (commonly invoked with
   `last_error()`) has been improved:
     - Display calls if present.